refactor: extract SiteInformationProps to higher parent directory
