+++ src/lib/map/index.ts
@@ -1,1 +1,8 @@
 export { MapVeuContainer as default } from './MapVeuContainer';
+
+export type SiteInformationProps = {
+  siteHomeUrl: string;
+  loginUrl: string;
+  siteName: string;
+  siteLogoSrc: string;
+};
